requirements.txt
beautifulsoup4>=4.9.3,<5.0.0
python-docx>=0.8.11
requests>=2.25.1
click>=8.0.0
typing-extensions>=4.0.0
colorlog>=6.7.0
Pillow>=10.0.0
yattag>=1.16.0
selenium>=4.0.0